pyproject.toml
[project]
name = "tree"
version = "1.0.0"
description = "Django project running fully on Docker"
requires-python = ">=3.12"

dependencies = [
    "Django==6.0.2",
    "django-filter==25.2",
    "djangorestframework==3.16.1",
    "Markdown==3.10.1",
    "psycopg2-binary==2.9.11",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-spectacular==0.29.0",
    "num2words==0.5.14",
    "pytz==2025.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
